refactor the dynamic validator to do FPSP and CSP separately and attempt to encapsulate some of the PDP changes
